% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-examples.R
\docType{data}
\name{bos}
\alias{bos}
\alias{bosms3}
\alias{bosms4}
\title{Bronchiolitis obliterans syndrome after lung transplants}
\format{
A data frame containing a sequence of observed or censored
transitions to the next stage of severity or death.  It is grouped
by patient and includes histories of 204 patients.  All patients
start in state 1 (no BOS) at six months after transplant, and may
subsequently develop BOS or die.
}
\source{
Papworth Hospital, U.K.
}
\description{
See \pkg{flexsurv} for details.
}
\references{
Heng. D. et al. (1998).  Bronchiolitis Obliterans Syndrome:
Incidence, Natural History, Prognosis, and Risk Factors.  Journal of Heart
and Lung Transplantation 17(12)1255--1263.
}
\keyword{datasets}
