% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data}
\alias{data}
\title{A fictional survival trial taken directly from survHE.}
\format{
An object of class \code{data.frame} with 367 rows and 8 columns.
}
\usage{
data
}
\description{
A dataset containing fictional data from a trial, where
the main outcome is in terms of time-to-event and
censoring indicator and with additional covariates.
}
\references{
\insertRef{Baio.2020}{expertsurv}
}
\keyword{datasets}
