% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper functions.R
\name{fit.models}
\alias{fit.models}
\title{Title}
\usage{
fit.models(formula = NULL, data, distr = NULL, method = "mle", exArgs, ...)
}
\arguments{
\item{formula}{}

\item{data}{}

\item{distr}{}

\item{method}{}

\item{...}{}
}
\value{

}
\description{
Title
}
