% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper functions.R
\name{get_quant_val}
\alias{get_quant_val}
\title{Title}
\usage{
get_quant_val(
  dist,
  param1,
  param2,
  param3 = NULL,
  probs = seq(0.01, 0.98, by = 0.01)
)
}
\arguments{
\item{dist}{}

\item{param1}{}

\item{param2}{}

\item{probs}{}
}
\value{

}
\description{
Title
}
