% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper functions.R
\name{makePoolPlot}
\alias{makePoolPlot}
\title{Title}
\usage{
makePoolPlot(
  fit,
  xl,
  xu,
  d = "best",
  w = 1,
  lwd = 1,
  xlab = "x",
  ylab = expression(f[X](x)),
  legend_full = TRUE,
  ql = NULL,
  qu = NULL,
  nx = 200,
  addquantile = FALSE,
  fs = 12,
  expertnames = NULL
)
}
\arguments{
\item{fit}{}

\item{xl}{}

\item{xu}{}

\item{d}{}

\item{w}{}

\item{lwd}{}

\item{xlab}{}

\item{ylab}{}

\item{legend_full}{}

\item{ql}{}

\item{qu}{}

\item{nx}{}

\item{addquantile}{}

\item{fs}{}

\item{expertnames}{}
}
\value{

}
\description{
Title
}
