% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-examples.R
\docType{data}
\name{bc}
\alias{bc}
\title{Breast cancer survival data}
\format{
An object of class \code{data.frame} with 686 rows and 4 columns.
}
\usage{
bc
}
\description{
Survival times of 686 patients with primary node positive breast cancer. See \pkg{flexsurv} for details.
}
\keyword{datasets}
