% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-examples.R
\docType{data}
\name{compiled_models_saved}
\alias{compiled_models_saved}
\title{Object with Compiled Stan Code}
\source{
Generated from \link[=compile_stan]{compile_stan}
}
\description{
Pre-compiled stan models that can be generated by the \link[=compile_stan]{compile_stan} function. These compiled stan models can be supplied to the \code{fit.models.expert} function with the argument \code{compile_mods}. This greatly facilitates the estimation of the parametric survival models run by stan.
}
\keyword{datasets}
