% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survHE_funcs.R
\name{compute_ICs_stan}
\alias{compute_ICs_stan}
\title{Helper function to compute the information criteria statistics
when using bayes as the inferential engine. 'rstan' does not do
DIC automatically and AIC/BIC are also not standard for Bayesian
models, so can compute them post-hoc by manipulating the
likelihood functions.}
\usage{
compute_ICs_stan(model, distr3, data.stan)
}
\arguments{
\item{model}{The 'rstan' object with the model fit}

\item{distr3}{The 'rstan' object with the model fit}
}
\value{
\item{list}{A list containing the modified name of the
distribution, the acronym (3-letters abbreviation), or the
labels (humane-readable name)}.
}
\description{
Helper function to compute the information criteria statistics
when using bayes as the inferential engine. 'rstan' does not do
DIC automatically and AIC/BIC are also not standard for Bayesian
models, so can compute them post-hoc by manipulating the
likelihood functions.
}
\note{
Something will go here
}
\references{
Baio (2020). survHE
}
\seealso{
fit.models
}
\author{
Gianluca Baio
}
\keyword{Approximation}
\keyword{Bayesian}
\keyword{Carlo}
\keyword{Hamiltonian}
\keyword{Integrated}
\keyword{Laplace}
\keyword{Monte}
\keyword{Nested}
\keyword{Parametric}
\keyword{inference}
\keyword{models}
\keyword{survival}
\keyword{via}
