% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_functions.R
\name{cred_int}
\alias{cred_int}
\title{Credible interval for pooled distribution}
\usage{
cred_int(plt_obj, val = "linear pool", interval = c(0.025, 0.975))
}
\arguments{
\item{plt_obj}{A plot object from \code{plot_expert_opinion}.}

\item{val}{The name of the opinion for which the interval will be generated.}

\item{interval}{A vector of the upper and lower probabilities. Default is the standard 95\% interval}
}
\value{

}
\description{
Returns the interval based on defined quantiles.
The approach used only provides an approximate (although quite accurate) integral.
}
\examples{
\dontrun{
cred_int(plot_opinion1,val = "linear pool", interval = c(0.025, 0.975))
}


}
