% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{elicit_surv}
\alias{elicit_surv}
\title{Elicit survival judgements interactively and estimate survival models}
\usage{
elicit_surv(compile_mods = compiled_models_saved)
}
\arguments{
\item{compile_mods}{list of compiled stan models generated by \code{compile_stan()}. Supplying the compiled stan models will greatly speed up the computation of the Bayesian analysis (otherwise each time a stan model is run it will be compiled (and not reused between runs)).}
}
\description{
Opens up a web browser (using the shiny package), from which you can specify
judgements and fit distributions for multiple timepoints and experts.
Plots of the fitted density functions are provided overlayed on the survival data (where appropriate).
}
\details{
Once the elicitation is complete the analysis can be run.
Click "Download R objects" to download the \code{expertsurv} object generated from the analysis.
Click "Download report" to generate a report including plots and parameter values for the parametric survival models.

For detailed instructions use \code{browseVignettes("expertsurv")}
}
\examples{
\dontrun{
library("dplyr")
library("ggplot2")
elicit_surv()
}
}
\author{
Philip Cooney \href{mailto:phcooney@tcd.ie}{phcooney@tcd.ie}
}
