% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expertsurv-package.R
\docType{package}
\name{expertsurv-package}
\alias{expertsurv-package}
\alias{expertsurv}
\title{Survival Analysis in Health Economic Evaluation}
\description{
Contains a suite of functions to perform survival analysis with the aim of
aiding in health economic modelling (extrapolation, model checking and PSA)
}
\details{
\tabular{ll}{ Package: \tab survHE\cr Type: \tab Package\cr Version: \tab
1.1.2\cr Date: \tab 2023-01-27\cr License: \tab GPL2 \cr LazyLoad: \tab
yes\cr } Contains a suite of functions to perform survival analysis with the
aim of aiding in health economic modelling (extrapolation, model checking
and PSA)
}
\examples{
\dontrun{ 
# Loads some survival data
data(bc)
# Fits a parametric model
m <- fit.models(formula=Surv(recyrs,censrec)~group,data=bc,
    distr="exp",method="mle")
# Print output in tabular format
print(m)
# Visualise output in terms of survival curves
plot(m)
}
}
\references{
P Cooney (2023). expertsurv: Incorporating expert opinion into parametric survival models
}
\author{
Philip Cooney

Maintainer: Philip Cooney
}
\keyword{Economic}
\keyword{Evaluation}
\keyword{Health}
\keyword{Modelling}
\keyword{Survival}
