% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexsurv-Integrate.R
\name{flexsurvreg}
\alias{flexsurvreg}
\title{Title}
\usage{
flexsurvreg(
  formula,
  anc = NULL,
  data,
  weights,
  bhazard,
  rtrunc,
  subset,
  na.action,
  dist,
  inits,
  fixedpars = NULL,
  dfns = NULL,
  aux = NULL,
  cl = 0.95,
  integ.opts = NULL,
  sr.control = survreg.control(),
  hessian = TRUE,
  hess.control = NULL,
  expert_opinion = NULL,
  ...
)
}
\arguments{
\item{formula}{}

\item{anc}{}

\item{data}{}

\item{weights}{}

\item{bhazard}{}

\item{rtrunc}{}

\item{subset}{}

\item{na.action}{}

\item{dist}{}

\item{inits}{}

\item{fixedpars}{}

\item{dfns}{}

\item{aux}{}

\item{cl}{}

\item{integ.opts}{}

\item{sr.control}{}

\item{hessian}{}

\item{hess.control}{}

\item{expert_opinion}{}

\item{...}{}
}
\value{

}
\description{
Title
}
